Consoldiate validator utils
