+++ packages/ui/src/helpers/accountSettings/index.ts
@@ -1,2 +1,6 @@
 export * from './utils';
-export * from './validators';
+export {
+  getPasswordConfirmationValidator,
+  getPasswordDefaultValidators,
+  runFieldValidators,
+} from './validator';
