+++ packages/ui/src/helpers/accountSettings/validator.ts
@@ -1,12 +1,14 @@
 import { Amplify } from 'aws-amplify';
-import { hasSpecialChars } from '../../../helpers/authenticator';
+import { hasSpecialChars } from '../authenticator';
 import {
   ValidatorSpec,
   PasswordSettings,
   PasswordRequirement,
-} from '../../../types';
+  ValidationMode,
+  InputEventType,
+} from '../../types';
 
-/** Gets password setting from Amplify configuration */
+// gets password requirement from Amplify.configure data
 export const getPasswordRequirement = (): PasswordRequirement => {
   // need to cast to any because `Amplify.configure()` isn't typed properly
   const config = Amplify.configure() as any;
@@ -31,7 +33,7 @@ export const getPasswordRequirement = (): PasswordRequirement => {
   };
 };
 
-export const getMinLengthValidator = (minLength: number): ValidatorSpec => ({
+export const getHasMinLength = (minLength: number): ValidatorSpec => ({
   validationMode: 'onTouched',
   validator: (field) => field.length >= minLength,
   message: `Password must have at least ${minLength} characters`,
@@ -61,7 +63,7 @@ export const hasSpecialChar: ValidatorSpec = {
   message: 'Password must have special characters',
 };
 
-export const getDefaultPasswordValidators = (): ValidatorSpec[] => {
+export const getPasswordDefaultValidators = (): ValidatorSpec[] => {
   const requirement = getPasswordRequirement();
   if (!requirement) return [];
 
@@ -76,7 +78,7 @@ export const getDefaultPasswordValidators = (): ValidatorSpec[] => {
   } = requirement;
 
   if (minLength) {
-    validators.push(getMinLengthValidator(minLength));
+    validators.push(getHasMinLength(minLength));
   }
 
   if (needsLowerCase) {
@@ -93,4 +95,70 @@ export const getDefaultPasswordValidators = (): ValidatorSpec[] => {
   }
 
   return validators;
+};
+
+export const getPasswordConfirmationValidator = (
+  password: string
+): ValidatorSpec => {
+  return {
+    validationMode: 'onTouched',
+    validator: (confirmPassword) => password === confirmPassword,
+    message: 'Your passwords must match',
+  };
+};
+
+/*
+ * `shouldValidate` determines whether validator should be run, based on validation mode,
+ * input event type, and whether it has been blurred yet.
+ */
+export const shouldValidate = ({
+  validationMode,
+  eventType,
+  hasBlurred,
+}: {
+  validationMode: ValidationMode;
+  eventType: InputEventType;
+  hasBlurred: boolean;
+}) => {
+  switch (validationMode) {
+    case 'onBlur': {
+      // only run validator on blur event
+      return eventType === 'blur';
+    }
+    case 'onChange': {
+      // only run validator on change event
+      return eventType === 'change';
+    }
+    case 'onTouched': {
+      /**
+       * run validator on first blur event, and then every subsequent
+       * blur/change event.
+       */
+      return eventType === 'blur' || hasBlurred;
+    }
+  }
+};
+
+// `runFieldValidator` runs all validators, and returns error messages.
+export const runFieldValidators = ({
+  value,
+  validators,
+  eventType,
+  hasBlurred,
+}: {
+  value: string;
+  validators: ValidatorSpec[];
+  eventType: InputEventType;
+  hasBlurred: boolean;
+}): string[] => {
+  if (!value) return [];
+
+  return validators.reduce((prevErrors, validatorSpec) => {
+    const { validator, validationMode, message } = validatorSpec;
+
+    if (shouldValidate({ validationMode, eventType, hasBlurred })) {
+      const hasError = !validator(value);
+      return hasError ? [...prevErrors, message] : prevErrors;
+    }
+  }, []);
 };
