--- packages/ui/src/helpers/accountSettings/validators/confirmPassword.ts
@@ -1,11 +1,0 @@
-import { ValidatorSpec } from '../../../types';
-
-export const getConfirmPasswordValidator = (
-  password: string
-): ValidatorSpec => {
-  return {
-    validationMode: 'onTouched',
-    validator: (confirmPassword) => password === confirmPassword,
-    message: 'Your passwords must match',
-  };
-};
