--- packages/ui/src/helpers/accountSettings/validators/index.ts
@@ -1,3 +1,0 @@
-export { getConfirmPasswordValidator } from './confirmPassword';
-export { getDefaultPasswordValidators } from './password';
-export { runFieldValidators } from './util';
